fix: MLflow artifact serving + dtype cast + registro de modelos existentes
- Dockerfile.mlflow: --artifacts-destination reemplaza --default-artifact-root
  para activar proxy HTTP en MLflow 3.x; allowed-hosts amplía localhost:5001
  y 127.0.0.1:5001 para acceso desde host macOS
- docker-compose.yml: monta mlflow_artifacts volume en el contenedor api
  para escritura directa al artifact store compartido
- predict.py: .astype("float64") en _feature_record_to_df y batch para evitar
  error de schema enforcement de MLflow pyfunc (int64 → float64)
- register_existing.py: script nuevo para registrar modelos joblib existentes
  en MLflow Registry sin reentrenar; usa transition_model_version_stage a
  Production para compatibilidad con dependencies.py

Verificado end-to-end: /predict devuelve churn_prob=0.017 con model_version=1

+++ src/api/routes/predict.py
@@ -29,7 +29,7 @@ router = APIRouter(prefix="/predict", tags=["predict"])
 def _feature_record_to_df(features) -> pd.DataFrame:
     """Convierte un FeatureRecord a DataFrame de una fila con las columnas correctas."""
     row = {k: getattr(features, k) for k in FEATURE_COLS_V5}
-    return pd.DataFrame([row])[FEATURE_COLS_V5]
+    return pd.DataFrame([row])[FEATURE_COLS_V5].astype("float64")
 
 
 def _build_response(
@@ -99,7 +99,7 @@ async def predict_batch(
         {k: getattr(u, k) for k in FEATURE_COLS_V5}
         for u in request.users
     ]
-    df     = pd.DataFrame(rows)[FEATURE_COLS_V5]
+    df     = pd.DataFrame(rows)[FEATURE_COLS_V5].astype("float64")
     probas = model_manager.predict(df)
     now    = datetime.utcnow()
     period = now.strftime("%Y-%m")
